requirements.txt
asgiref==3.2.7
Django==3.0.6
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
pytz==2020.1
sqlparse==0.3.1
